Move more structure's parsing to custom files
